+++ pkgs/test/test/runner/json_reporter_test.dart
@@ -5,16 +5,14 @@
 @TestOn('vm')
 
 import 'dart:async';
-import 'dart:convert';
 
 import 'package:path/path.dart' as p;
-
 import 'package:test_descriptor/test_descriptor.dart' as d;
 
-import 'package:test_core/src/runner/version.dart';
 import 'package:test/test.dart';
 
 import '../io.dart';
+import 'json_reporter_utils.dart';
 
 void main() {
   test('runs several successful tests and reports when each completes', () {
@@ -24,20 +22,20 @@ void main() {
       test('success 3', () {});
     ''', [
       [
-        _suite(0),
-        _testStart(1, 'loading test.dart', groupIDs: []),
-        _testDone(1, hidden: true),
+        suiteJson(0),
+        testStartJson(1, 'loading test.dart', groupIDs: []),
+        testDoneJson(1, hidden: true),
       ],
       [
-        _group(2, testCount: 3),
-        _testStart(3, 'success 1', line: 6, column: 7),
-        _testDone(3),
-        _testStart(4, 'success 2', line: 7, column: 7),
-        _testDone(4),
-        _testStart(5, 'success 3', line: 8, column: 7),
-        _testDone(5),
+        groupJson(2, testCount: 3),
+        testStartJson(3, 'success 1', line: 6, column: 7),
+        testDoneJson(3),
+        testStartJson(4, 'success 2', line: 7, column: 7),
+        testDoneJson(4),
+        testStartJson(5, 'success 3', line: 8, column: 7),
+        testDoneJson(5),
       ]
-    ], _done());
+    ], doneJson());
   });
 
   test('runs several failing tests and reports when each fails', () {
@@ -47,23 +45,23 @@ void main() {
       test('failure 3', () => throw TestFailure('oh no'));
     ''', [
       [
-        _suite(0),
-        _testStart(1, 'loading test.dart', groupIDs: []),
-        _testDone(1, hidden: true),
+        suiteJson(0),
+        testStartJson(1, 'loading test.dart', groupIDs: []),
+        testDoneJson(1, hidden: true),
       ],
       [
-        _group(2, testCount: 3),
-        _testStart(3, 'failure 1', line: 6, column: 7),
-        _error(3, 'oh no', isFailure: true),
-        _testDone(3, result: 'failure'),
-        _testStart(4, 'failure 2', line: 7, column: 7),
-        _error(4, 'oh no', isFailure: true),
-        _testDone(4, result: 'failure'),
-        _testStart(5, 'failure 3', line: 8, column: 7),
-        _error(5, 'oh no', isFailure: true),
-        _testDone(5, result: 'failure'),
+        groupJson(2, testCount: 3),
+        testStartJson(3, 'failure 1', line: 6, column: 7),
+        errorJson(3, 'oh no', isFailure: true),
+        testDoneJson(3, result: 'failure'),
+        testStartJson(4, 'failure 2', line: 7, column: 7),
+        errorJson(4, 'oh no', isFailure: true),
+        testDoneJson(4, result: 'failure'),
+        testStartJson(5, 'failure 3', line: 8, column: 7),
+        errorJson(5, 'oh no', isFailure: true),
+        testDoneJson(5, result: 'failure'),
       ]
-    ], _done(success: false));
+    ], doneJson(success: false));
   });
 
   test('includes the full stack trace with --verbose-trace', () async {
@@ -91,24 +89,24 @@ void main() {
       test('success 2', () {});
     ''', [
       [
-        _suite(0),
-        _testStart(1, 'loading test.dart', groupIDs: []),
-        _testDone(1, hidden: true),
+        suiteJson(0),
+        testStartJson(1, 'loading test.dart', groupIDs: []),
+        testDoneJson(1, hidden: true),
       ],
       [
-        _group(2, testCount: 4),
-        _testStart(3, 'failure 1', line: 6, column: 7),
-        _error(3, 'oh no', isFailure: true),
-        _testDone(3, result: 'failure'),
-        _testStart(4, 'success 1', line: 7, column: 7),
-        _testDone(4),
-        _testStart(5, 'failure 2', line: 8, column: 7),
-        _error(5, 'oh no', isFailure: true),
-        _testDone(5, result: 'failure'),
-        _testStart(6, 'success 2', line: 9, column: 7),
-        _testDone(6),
+        groupJson(2, testCount: 4),
+        testStartJson(3, 'failure 1', line: 6, column: 7),
+        errorJson(3, 'oh no', isFailure: true),
+        testDoneJson(3, result: 'failure'),
+        testStartJson(4, 'success 1', line: 7, column: 7),
+        testDoneJson(4),
+        testStartJson(5, 'failure 2', line: 8, column: 7),
+        errorJson(5, 'oh no', isFailure: true),
+        testDoneJson(5, result: 'failure'),
+        testStartJson(6, 'success 2', line: 9, column: 7),
+        testDoneJson(6),
       ]
-    ], _done(success: false));
+    ], doneJson(success: false));
   });
 
   test('gracefully handles multiple test failures in a row', () {
@@ -125,21 +123,21 @@ void main() {
       test('wait', () => completer.future);
     ''', [
       [
-        _suite(0),
-        _testStart(1, 'loading test.dart', groupIDs: []),
-        _testDone(1, hidden: true),
+        suiteJson(0),
+        testStartJson(1, 'loading test.dart', groupIDs: []),
+        testDoneJson(1, hidden: true),
       ],
       [
-        _group(2, testCount: 2),
-        _testStart(3, 'failures', line: 9, column: 7),
-        _error(3, 'first error'),
-        _error(3, 'second error'),
-        _error(3, 'third error'),
-        _testDone(3, result: 'error'),
-        _testStart(4, 'wait', line: 15, column: 7),
-        _testDone(4),
+        groupJson(2, testCount: 2),
+        testStartJson(3, 'failures', line: 9, column: 7),
+        errorJson(3, 'first error'),
+        errorJson(3, 'second error'),
+        errorJson(3, 'third error'),
+        testDoneJson(3, result: 'error'),
+        testStartJson(4, 'wait', line: 15, column: 7),
+        testDoneJson(4),
       ]
-    ], _done(success: false));
+    ], doneJson(success: false));
   });
 
   test('gracefully handles a test failing after completion', () {
@@ -161,24 +159,24 @@ void main() {
       });
     ''', [
       [
-        _suite(0),
-        _testStart(1, 'loading test.dart', groupIDs: []),
-        _testDone(1, hidden: true),
+        suiteJson(0),
+        testStartJson(1, 'loading test.dart', groupIDs: []),
+        testDoneJson(1, hidden: true),
       ],
       [
-        _group(2, testCount: 2),
-        _testStart(3, 'failure', line: 11, column: 7),
-        _testDone(3),
-        _testStart(4, 'wait', line: 17, column: 7),
-        _error(3, 'oh no'),
-        _error(
+        groupJson(2, testCount: 2),
+        testStartJson(3, 'failure', line: 11, column: 7),
+        testDoneJson(3),
+        testStartJson(4, 'wait', line: 17, column: 7),
+        errorJson(3, 'oh no'),
+        errorJson(
             3,
             'This test failed after it had already completed. Make sure to '
             'use [expectAsync]\n'
             'or the [completes] matcher when testing async code.'),
-        _testDone(4),
+        testDoneJson(4),
       ]
-    ], _done(success: false));
+    ], doneJson(success: false));
   });
 
   test('reports each test in its proper groups', () {
@@ -195,25 +193,27 @@ void main() {
       });
     ''', [
       [
-        _suite(0),
-        _testStart(1, 'loading test.dart', groupIDs: []),
-        _testDone(1, hidden: true),
+        suiteJson(0),
+        testStartJson(1, 'loading test.dart', groupIDs: []),
+        testDoneJson(1, hidden: true),
       ],
       [
-        _group(2, testCount: 3),
-        _group(3,
+        groupJson(2, testCount: 3),
+        groupJson(3,
             name: 'group 1', parentID: 2, testCount: 3, line: 6, column: 7),
-        _group(4, name: 'group 1 .2', parentID: 3, line: 7, column: 9),
-        _group(5, name: 'group 1 .2 .3', parentID: 4, line: 8, column: 11),
-        _testStart(6, 'group 1 .2 .3 success',
+        groupJson(4, name: 'group 1 .2', parentID: 3, line: 7, column: 9),
+        groupJson(5, name: 'group 1 .2 .3', parentID: 4, line: 8, column: 11),
+        testStartJson(6, 'group 1 .2 .3 success',
             groupIDs: [2, 3, 4, 5], line: 9, column: 13),
-        _testDone(6),
-        _testStart(7, 'group 1 success', groupIDs: [2, 3], line: 13, column: 9),
-        _testDone(7),
-        _testStart(8, 'group 1 success', groupIDs: [2, 3], line: 14, column: 9),
-        _testDone(8),
+        testDoneJson(6),
+        testStartJson(7, 'group 1 success',
+            groupIDs: [2, 3], line: 13, column: 9),
+        testDoneJson(7),
+        testStartJson(8, 'group 1 success',
+            groupIDs: [2, 3], line: 14, column: 9),
+        testDoneJson(8),
       ]
-    ], _done());
+    ], doneJson());
   });
 
   group('print:', () {
@@ -227,20 +227,20 @@ void main() {
         });
       ''', [
         [
-          _suite(0),
-          _testStart(1, 'loading test.dart', groupIDs: []),
-          _testDone(1, hidden: true),
+          suiteJson(0),
+          testStartJson(1, 'loading test.dart', groupIDs: []),
+          testDoneJson(1, hidden: true),
         ],
         [
-          _group(2),
-          _testStart(3, 'test', line: 6, column: 9),
-          _print(3, 'one'),
-          _print(3, 'two'),
-          _print(3, 'three'),
-          _print(3, 'four'),
-          _testDone(3),
+          groupJson(2),
+          testStartJson(3, 'test', line: 6, column: 9),
+          printJson(3, 'one'),
+          printJson(3, 'two'),
+          printJson(3, 'three'),
+          printJson(3, 'four'),
+          testDoneJson(3),
         ]
-      ], _done());
+      ], doneJson());
     });
 
     test('handles a print after the test completes', () {
@@ -265,22 +265,22 @@ void main() {
         });
       ''', [
         [
-          _suite(0),
-          _testStart(1, 'loading test.dart', groupIDs: []),
-          _testDone(1, hidden: true),
+          suiteJson(0),
+          testStartJson(1, 'loading test.dart', groupIDs: []),
+          testDoneJson(1, hidden: true),
         ],
         [
-          _group(2, testCount: 2),
-          _testStart(3, 'test', line: 10, column: 9),
-          _testDone(3),
-          _testStart(4, 'wait', line: 20, column: 9),
-          _print(3, 'one'),
-          _print(3, 'two'),
-          _print(3, 'three'),
-          _print(3, 'four'),
-          _testDone(4),
+          groupJson(2, testCount: 2),
+          testStartJson(3, 'test', line: 10, column: 9),
+          testDoneJson(3),
+          testStartJson(4, 'wait', line: 20, column: 9),
+          printJson(3, 'one'),
+          printJson(3, 'two'),
+          printJson(3, 'three'),
+          printJson(3, 'four'),
+          testDoneJson(4),
         ]
-      ], _done());
+      ], doneJson());
     });
 
     test('interleaves prints and errors', () {
@@ -309,26 +309,26 @@ void main() {
         test('wait', () => completer.future);
       ''', [
         [
-          _suite(0),
-          _testStart(1, 'loading test.dart', groupIDs: []),
-          _testDone(1, hidden: true),
+          suiteJson(0),
+          testStartJson(1, 'loading test.dart', groupIDs: []),
+          testDoneJson(1, hidden: true),
         ],
         [
-          _group(2, testCount: 2),
-          _testStart(3, 'test', line: 9, column: 9),
-          _print(3, 'one'),
-          _print(3, 'two'),
-          _error(3, 'first error'),
-          _print(3, 'three'),
-          _print(3, 'four'),
-          _error(3, 'second error'),
-          _print(3, 'five'),
-          _print(3, 'six'),
-          _testDone(3, result: 'error'),
-          _testStart(4, 'wait', line: 27, column: 9),
-          _testDone(4),
+          groupJson(2, testCount: 2),
+          testStartJson(3, 'test', line: 9, column: 9),
+          printJson(3, 'one'),
+          printJson(3, 'two'),
+          errorJson(3, 'first error'),
+          printJson(3, 'three'),
+          printJson(3, 'four'),
+          errorJson(3, 'second error'),
+          printJson(3, 'five'),
+          printJson(3, 'six'),
+          testDoneJson(3, result: 'error'),
+          testStartJson(4, 'wait', line: 27, column: 9),
+          testDoneJson(4),
         ]
-      ], _done(success: false));
+      ], doneJson(success: false));
     });
   });
 
@@ -340,20 +340,20 @@ void main() {
         test('skip 3', () {}, skip: true);
       ''', [
         [
-          _suite(0),
-          _testStart(1, 'loading test.dart', groupIDs: []),
-          _testDone(1, hidden: true),
+          suiteJson(0),
+          testStartJson(1, 'loading test.dart', groupIDs: []),
+          testDoneJson(1, hidden: true),
         ],
         [
-          _group(2, testCount: 3),
-          _testStart(3, 'skip 1', skip: true, line: 6, column: 9),
-          _testDone(3, skipped: true),
-          _testStart(4, 'skip 2', skip: true, line: 7, column: 9),
-          _testDone(4, skipped: true),
-          _testStart(5, 'skip 3', skip: true, line: 8, column: 9),
-          _testDone(5, skipped: true),
+          groupJson(2, testCount: 3),
+          testStartJson(3, 'skip 1', skip: true, line: 6, column: 9),
+          testDoneJson(3, skipped: true),
+          testStartJson(4, 'skip 2', skip: true, line: 7, column: 9),
+          testDoneJson(4, skipped: true),
+          testStartJson(5, 'skip 3', skip: true, line: 8, column: 9),
+          testDoneJson(5, skipped: true),
         ]
-      ], _done());
+      ], doneJson());
     });
 
     test('reports skipped groups', () {
@@ -365,30 +365,30 @@ void main() {
         }, skip: true);
       ''', [
         [
-          _suite(0),
-          _testStart(1, 'loading test.dart', groupIDs: []),
-          _testDone(1, hidden: true),
+          suiteJson(0),
+          testStartJson(1, 'loading test.dart', groupIDs: []),
+          testDoneJson(1, hidden: true),
         ],
         [
-          _group(2, testCount: 3),
-          _group(3,
+          groupJson(2, testCount: 3),
+          groupJson(3,
               name: 'skip',
               parentID: 2,
               skip: true,
               testCount: 3,
               line: 6,
               column: 9),
-          _testStart(4, 'skip success 1',
+          testStartJson(4, 'skip success 1',
               groupIDs: [2, 3], skip: true, line: 7, column: 11),
-          _testDone(4, skipped: true),
-          _testStart(5, 'skip success 2',
+          testDoneJson(4, skipped: true),
+          testStartJson(5, 'skip success 2',
               groupIDs: [2, 3], skip: true, line: 8, column: 11),
-          _testDone(5, skipped: true),
-          _testStart(6, 'skip success 3',
+          testDoneJson(5, skipped: true),
+          testStartJson(6, 'skip success 3',
               groupIDs: [2, 3], skip: true, line: 9, column: 11),
-          _testDone(6, skipped: true),
+          testDoneJson(6, skipped: true),
         ]
-      ], _done());
+      ], doneJson());
     });
 
     test('reports the skip reason if available', () {
@@ -397,20 +397,20 @@ void main() {
         test('skip 2', () {}, skip: 'or another');
       ''', [
         [
-          _suite(0),
-          _testStart(1, 'loading test.dart', groupIDs: []),
-          _testDone(1, hidden: true),
+          suiteJson(0),
+          testStartJson(1, 'loading test.dart', groupIDs: []),
+          testDoneJson(1, hidden: true),
         ],
         [
-          _group(2, testCount: 2),
-          _testStart(3, 'skip 1', skip: 'some reason', line: 6, column: 9),
-          _print(3, 'Skip: some reason', type: 'skip'),
-          _testDone(3, skipped: true),
-          _testStart(4, 'skip 2', skip: 'or another', line: 7, column: 9),
-          _print(4, 'Skip: or another', type: 'skip'),
-          _testDone(4, skipped: true),
+          groupJson(2, testCount: 2),
+          testStartJson(3, 'skip 1', skip: 'some reason', line: 6, column: 9),
+          printJson(3, 'Skip: some reason', type: 'skip'),
+          testDoneJson(3, skipped: true),
+          testStartJson(4, 'skip 2', skip: 'or another', line: 7, column: 9),
+          printJson(4, 'Skip: or another', type: 'skip'),
+          testDoneJson(4, skipped: true),
         ]
-      ], _done());
+      ], doneJson());
     });
 
     test('runs skipped tests with --run-skipped', () {
@@ -421,19 +421,19 @@ void main() {
       ''',
           [
             [
-              _suite(0),
-              _testStart(1, 'loading test.dart', groupIDs: []),
-              _testDone(1, hidden: true),
+              suiteJson(0),
+              testStartJson(1, 'loading test.dart', groupIDs: []),
+              testDoneJson(1, hidden: true),
             ],
             [
-              _group(2, testCount: 2),
-              _testStart(3, 'skip 1', line: 6, column: 9),
-              _testDone(3),
-              _testStart(4, 'skip 2', line: 7, column: 9),
-              _testDone(4),
+              groupJson(2, testCount: 2),
+              testStartJson(3, 'skip 1', line: 6, column: 9),
+              testDoneJson(3),
+              testStartJson(4, 'skip 2', line: 7, column: 9),
+              testDoneJson(4),
             ]
           ],
-          _done(),
+          doneJson(),
           args: ['--run-skipped']);
     });
   });
@@ -447,20 +447,20 @@ void main() {
         test('success', () {});
       ''', [
         [
-          _suite(0),
-          _testStart(1, 'loading test.dart', groupIDs: []),
-          _testDone(1, hidden: true),
+          suiteJson(0),
+          testStartJson(1, 'loading test.dart', groupIDs: []),
+          testDoneJson(1, hidden: true),
         ],
         [
-          _group(2, testCount: 1),
-          _testStart(3, '(setUpAll)', line: 6, column: 9),
-          _testDone(3, hidden: true),
-          _testStart(4, 'success', line: 9, column: 9),
-          _testDone(4),
-          _testStart(5, '(tearDownAll)'),
-          _testDone(5, hidden: true),
+          groupJson(2, testCount: 1),
+          testStartJson(3, '(setUpAll)', line: 6, column: 9),
+          testDoneJson(3, hidden: true),
+          testStartJson(4, 'success', line: 9, column: 9),
+          testDoneJson(4),
+          testStartJson(5, '(tearDownAll)'),
+          testDoneJson(5, hidden: true),
         ]
-      ], _done());
+      ], doneJson());
     });
 
     test('the first call to tearDownAll()', () {
@@ -471,18 +471,18 @@ void main() {
         test('success', () {});
       ''', [
         [
-          _testStart(1, 'loading test.dart', groupIDs: []),
-          _testDone(1, hidden: true),
+          testStartJson(1, 'loading test.dart', groupIDs: []),
+          testDoneJson(1, hidden: true),
         ],
         [
-          _suite(0),
-          _group(2, testCount: 1),
-          _testStart(3, 'success', line: 9, column: 9),
-          _testDone(3),
-          _testStart(4, '(tearDownAll)', line: 6, column: 9),
-          _testDone(4, hidden: true),
+          suiteJson(0),
+          groupJson(2, testCount: 1),
+          testStartJson(3, 'success', line: 9, column: 9),
+          testDoneJson(3),
+          testStartJson(4, '(tearDownAll)', line: 6, column: 9),
+          testDoneJson(4, hidden: true),
         ]
-      ], _done());
+      ], doneJson());
     });
 
     test('a test compiled to JS', () {
@@ -492,18 +492,18 @@ void main() {
       ''',
           [
             [
-              _suite(0, platform: 'chrome'),
-              _testStart(1, 'compiling test.dart', groupIDs: []),
-              _print(1, startsWith('Compiled')),
-              _testDone(1, hidden: true),
+              suiteJson(0, platform: 'chrome'),
+              testStartJson(1, 'compiling test.dart', groupIDs: []),
+              printJson(1, startsWith('Compiled')),
+              testDoneJson(1, hidden: true),
             ],
             [
-              _group(2, testCount: 1),
-              _testStart(3, 'success', line: 6, column: 9),
-              _testDone(3),
+              groupJson(2, testCount: 1),
+              testStartJson(3, 'success', line: 6, column: 9),
+              testDoneJson(3),
             ]
           ],
-          _done(),
+          doneJson(),
           args: ['-p', 'chrome']);
     }, tags: ['chrome'], skip: 'https://github.com/dart-lang/test/issues/872');
 
@@ -515,25 +515,25 @@ void main() {
     ''',
           [
             [
-              _suite(0),
-              _testStart(1, 'loading test.dart', groupIDs: []),
-              _testDone(1, hidden: true),
+              suiteJson(0),
+              testStartJson(1, 'loading test.dart', groupIDs: []),
+              testDoneJson(1, hidden: true),
             ],
             [
-              _group(2, testCount: 2),
-              _testStart(3, 'success 1',
+              groupJson(2, testCount: 2),
+              testStartJson(3, 'success 1',
                   line: 3,
                   column: 50,
                   url: p.toUri(p.join(d.sandbox, 'common.dart')).toString(),
                   root_column: 7,
                   root_line: 7,
                   root_url: p.toUri(p.join(d.sandbox, 'test.dart')).toString()),
-              _testDone(3),
-              _testStart(4, 'success 2', line: 8, column: 7),
-              _testDone(4),
+              testDoneJson(3),
+              testStartJson(4, 'success 2', line: 8, column: 7),
+              testDoneJson(4),
             ]
           ],
-          _done(),
+          doneJson(),
           externalLibraries: {
             'common.dart': '''
 import 'package:test/test.dart';
@@ -553,18 +553,18 @@ void customTest(String name, Function testFn) => test(name, testFn);
     ''',
         [
           [
-            _suite(0, platform: 'chrome'),
-            _testStart(1, 'compiling test.dart', groupIDs: []),
-            _print(1, startsWith('Compiled')),
-            _testDone(1, hidden: true),
+            suiteJson(0, platform: 'chrome'),
+            testStartJson(1, 'compiling test.dart', groupIDs: []),
+            printJson(1, startsWith('Compiled')),
+            testDoneJson(1, hidden: true),
           ],
           [
-            _group(2, testCount: 1),
-            _testStart(3, 'success'),
-            _testDone(3),
+            groupJson(2, testCount: 1),
+            testStartJson(3, 'success'),
+            testDoneJson(3),
           ],
         ],
-        _done(),
+        doneJson(),
         args: ['-p', 'chrome', '--js-trace']);
   }, tags: ['chrome']);
 }
@@ -598,205 +598,5 @@ import 'package:test/test.dart';
   await test.shouldExit();
 
   var stdoutLines = await test.stdoutStream().toList();
-  // Ensure the output is of the same length, including start, done and all
-  // suites messages.
-  expect(stdoutLines.length, equals(expected.fold(3, (a, m) => a + m.length)),
-      reason: 'Expected $stdoutLines to match $expected.');
-
-  dynamic decodeLine(String l) =>
-      jsonDecode(l)..remove('time')..remove('stackTrace');
-
-  // Should contain all suites message.
-  expect(stdoutLines.map(decodeLine), containsAll([_allSuites()]));
-
-  // A single start event is emitted first.
-  final _start = {
-    'type': 'start',
-    'protocolVersion': '0.1.1',
-    'runnerVersion': testVersion,
-    'pid': test.pid
-  };
-  expect(decodeLine(stdoutLines.first), equals(_start));
-
-  // A single done event is emmited last.
-  expect(decodeLine(stdoutLines.last), equals(done));
-
-  for (var value in expected) {
-    expect(stdoutLines.map(decodeLine), containsAllInOrder(value));
-  }
-}
-
-/// Returns the event emitted by the JSON reporter providing information about
-/// all suites.
-///
-/// The [count] defaults to 1.
-Map _allSuites({int count}) {
-  return {'type': 'allSuites', 'count': count ?? 1};
-}
-
-/// Returns the event emitted by the JSON reporter indicating that a suite has
-/// begun running.
-///
-/// The [platform] defaults to `"vm"`, the [path] defaults to `"test.dart"`.
-Map _suite(int id, {String platform, String path}) {
-  return {
-    'type': 'suite',
-    'suite': {
-      'id': id,
-      'platform': platform ?? 'vm',
-      'path': path ?? 'test.dart'
-    }
-  };
-}
-
-/// Returns the event emitted by the JSON reporter indicating that a group has
-/// begun running.
-///
-/// If [skip] is `true`, the group is expected to be marked as skipped without a
-/// reason. If it's a [String], the group is expected to be marked as skipped
-/// with that reason.
-///
-/// The [testCount] parameter indicates the number of tests in the group. It
-/// defaults to 1.
-Map _group(int id,
-    {String name,
-    int suiteID,
-    int parentID,
-    skip,
-    int testCount,
-    int line,
-    int column}) {
-  if ((line == null) != (column == null)) {
-    throw ArgumentError(
-        'line and column must either both be null or both be passed');
-  }
-
-  return {
-    'type': 'group',
-    'group': {
-      'id': id,
-      'name': name,
-      'suiteID': suiteID ?? 0,
-      'parentID': parentID,
-      'metadata': _metadata(skip: skip),
-      'testCount': testCount ?? 1,
-      'line': line,
-      'column': column,
-      'url': line == null
-          ? null
-          : p.toUri(p.join(d.sandbox, 'test.dart')).toString()
-    }
-  };
-}
-
-/// Returns the event emitted by the JSON reporter indicating that a test has
-/// begun running.
-///
-/// If [parentIDs] is passed, it's the IDs of groups containing this test. If
-/// [skip] is `true`, the test is expected to be marked as skipped without a
-/// reason. If it's a [String], the test is expected to be marked as skipped
-/// with that reason.
-Map _testStart(int id, String name,
-    {int suiteID,
-    Iterable<int> groupIDs,
-    int line,
-    int column,
-    String url,
-    skip,
-    int root_line,
-    int root_column,
-    String root_url}) {
-  if ((line == null) != (column == null)) {
-    throw ArgumentError(
-        'line and column must either both be null or both be passed');
-  }
-
-  url ??=
-      line == null ? null : p.toUri(p.join(d.sandbox, 'test.dart')).toString();
-  var expected = {
-    'type': 'testStart',
-    'test': {
-      'id': id,
-      'name': name,
-      'suiteID': suiteID ?? 0,
-      'groupIDs': groupIDs ?? [2],
-      'metadata': _metadata(skip: skip),
-      'line': line,
-      'column': column,
-      'url': url,
-    }
-  };
-  var testObj = expected['test'] as Map<String, dynamic>;
-  if (root_line != null) {
-    testObj['root_line'] = root_line;
-  }
-  if (root_column != null) {
-    testObj['root_column'] = root_column;
-  }
-  if (root_url != null) {
-    testObj['root_url'] = root_url;
-  }
-  return expected;
-}
-
-/// Returns the event emitted by the JSON reporter indicating that a test
-/// printed [message].
-Matcher _print(int id, dynamic /*String|Matcher*/ message, {String type}) {
-  return allOf(
-    hasLength(4),
-    containsPair('type', 'print'),
-    containsPair('testID', id),
-    containsPair('message', message),
-    containsPair('messageType', type ?? 'print'),
-  );
-}
-
-/// Returns the event emitted by the JSON reporter indicating that a test
-/// emitted [error].
-///
-/// The [isFailure] parameter indicates whether the error was a [TestFailure] or
-/// not.
-Map _error(int id, String error, {bool isFailure = false}) {
-  return {
-    'type': 'error',
-    'testID': id,
-    'error': error,
-    'isFailure': isFailure
-  };
-}
-
-/// Returns the event emitted by the JSON reporter indicating that a test
-/// finished.
-///
-/// The [result] parameter indicates the result of the test. It defaults to
-/// `"success"`.
-///
-/// The [hidden] parameter indicates whether the test should not be displayed
-/// after finishing. The [skipped] parameter indicates whether the test was
-/// skipped.
-Map _testDone(int id,
-    {String result, bool hidden = false, bool skipped = false}) {
-  result ??= 'success';
-  return {
-    'type': 'testDone',
-    'testID': id,
-    'result': result,
-    'hidden': hidden,
-    'skipped': skipped
-  };
-}
-
-/// Returns the event emitted by the JSON reporter indicating that the entire
-/// run finished.
-Map _done({bool success = true}) => {'type': 'done', 'success': success};
-
-/// Returns the serialized metadata corresponding to [skip].
-Map _metadata({skip}) {
-  if (skip == true) {
-    return {'skip': true, 'skipReason': null};
-  } else if (skip is String) {
-    return {'skip': true, 'skipReason': skip};
-  } else {
-    return {'skip': false, 'skipReason': null};
-  }
+  return expectJsonReport(stdoutLines, test.pid, expected, done);
 }