+++ pkgs/test_core/CHANGELOG.md
@@ -1,6 +1,7 @@
-## 0.2.17-dev
+## 0.2.17
 
-* Internal cleanup.
+* Add `file_reporters` configuration option and `--file-reporter` CLI option to
+  allow specifying a separate reporter that writes to a file instead of stdout.
 
 ## 0.2.16